ai: Only go for the ball when its coming towards us
If the ball's velocity is heading towards the enemy paddle, then
start aiming for it as simply as we can. Otherwise, we need to return
to the homebase.

Due to some uggly jittering we've had to do some casting, which removes
most of it. We'll also need to add some wiggle tolerance to remove the
last of the enemy paddle jittering.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>