Rework paddle logic completely for obstacles
We now start the obstacle paddles in the dead-center-y of the play
arena, and give them opposing velocities. Additionally we equip the
paddles with two sensors, top and bottom, to let us know when we hit
an 'edge'.

Upon hitting any given edge, we automatically invert the velocity of
the obstacle, thus simplifying life dramatically.

The AI logic now deliberately skips any 'None' edge obstacles to
avoid chasing issues.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>