ai: Make our logic more robust, solve chasing issues
When scoring against the enemy, we never 'return to home' the entity as
it will keep chasing the last ball. As such, we'll now always return the
entity to home when there are no balls.

The logic has been made more robust in the AI controller - sharing some code
to assist with commonly used functionality.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>